working, need buffer for menu
