+++ game/source/gameMenu.c
@@ -1,15 +1,20 @@
 #include <stdint.h>
 #include "../include/gameMenu.h"
 #include "../../drivers/include/OLEDDriver.h"
-#include "../../drivers/include/ADCDriver.h"
-
-
-
+#include "../../drivers/include/userInputDriver.h"
+#include <util/delay.h>
 
 void menuLoop(struct Node* startNode){
   uint8_t selectedOption = 0;
+  JoystickDir currentDir;
+  currentDir = calculateJoystickDirection();
   struct Node* currentNode = startNode;
-  printNode(currentNode, 1);
+  while(1){
+    if (startNode != currentNode){
+
+    }
+    printNode(currentNode, selectedOption);
+  }
 }
 
 void printNode(struct Node* node, uint8_t selectedOption){
@@ -25,4 +30,5 @@ void printNode(struct Node* node, uint8_t selectedOption){
     OLED_print(node->options[i]);
     printf(node->options[i]);
   }
+  _delay_ms(500);
 }
