+++ toDo.txt
@@ -2,3 +2,5 @@
 * printf til skjerm (hvis du føler deg fancy)
 * Kalibreringsrutine til joystick og slider
 * Trigonometri joystickretningfunskjon
+* Minnebuffer til skjerm (slipper å tømme og skrive over alt hver gang)
+* logikk kun oppdaterer sjerm på de områdene som  er forandret
